Moved some things around to accommodate posting to the GET url.  (If you're looking at a form at "/a/b/c", then the form POSTs to "/a/b/c".)
The function that handled adding new posts from the admin is now in the AdminHandler.  All of the requests to the AdminHandler verify that a user is logged in, whereas the ContentHandler (my impression may be wrong) would handle incoming requests from places that would be inherently not login-secured, such as comments, trackbacks, pingbacks,  etc.

I built the code that we had discussed in the dev thread to handle this change:
http://groups.google.com/group/habari-dev/browse_thread/thread/df1735a3fd19f4ca/298a877696b4979d

I added success messages that work, so you can use these as a prototype for the future.  The redirects all use Util::redirect().  The URLs used are all obtained via URL::get().  In order to post to the correct location, I removed the URL rule that handled posts in ContentHandler.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@130 653ae4dd-d31e-0410-96ef-6bf7bf53c507