Ah, the magic of inheritance.
Because ContentHandler extends ActionHandler, it automatically dispatches incoming requests to a member with the same name as the action (third array element) in the URL rule.

By crafting the URL rule with this in mind, you can eliminate two functions from ContentHandler that already exist in the base class, while making the URL for posting comments i18n-friendly, extendable by plugins, or random for each Habari install.  (This is good for preventing spam from bots that expect the comment entrypoint to be the same (ie- "/wp-post-comment.php") on every blog it posts to.)

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@131 653ae4dd-d31e-0410-96ef-6bf7bf53c507