Fixed the tag table schema to allow multiple tags on one post.
Post::get_tags() must always return the array of tags attached to the post.

Added formating function (in Util class, although this should move to a Format class) to convert array of tags into an HTML list with separators, as in prior revision.  Modified the template to use this function on the post tag array, ala echo Utils::tag_and_list($post->tags)

Changed the method of getting a comment count so that the comments are properly memory-cached and the count is correctly returned every time, not just the first time.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@136 653ae4dd-d31e-0410-96ef-6bf7bf53c507