* added support for pagination in Comments::get() and Posts::get() by use of LIMIT and OFFSET parameters.
* added support to Posts::get() to specify the DB function to execute.  
Add 'fetch_fn' to the list of parameters to specify:
   $posts = Posts::get(array('slug' => 'foo', 'fetch_fn' => 'get_row'));
This allows Posts::get() to return an array, or a single Post object 
(primarily for use with Post::get(), but also for use with the various 
count_* functions)

* Post::get() now calls Posts::get() with 'fetch_fn' => 'get_row'

* Comments::get() and Posts::get() should both return either an array
(possibly empty) of their respective objects, or a single value (for use
with the count_* functions in each class).  Removed the condition that
would return boolean FALSE

* updated admin/options to prompt for a default value for pagination.

* updated admin/content to override default pagination.

* updated installer to set default pagination value to 5.

* fixed a few "undefined index" warnings in adminhandler, post, and url 
classes.

* changed case of "id" value in user class to match case used by 
database column.



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@159 653ae4dd-d31e-0410-96ef-6bf7bf53c507