Big commit, adding some user management features.
* removed all instances of nickname.  Habari users now only have a 
username.  If this decision stays, we should probably rename the 
"username" column to simply "name":  $user->username is a bit redundant.

* added a Users admin menu item, and associated page.  This lists all 
the users on the system, and displays some basic info about each 
account.  Provides a link to each user's profile page.

* added a form to create new users on the /admin/users page. (ticket 
#15) Needs more error checking, probably.

* when viewing another user's profile at /admin/user/foo, there is a 
button to delete that user.  (ticket #43).  You may not currently delete 
your own user account.  Deleting a user does not modify the post table 
at all: opening a new ticket to track this defect.

We need some form of authorization, to determine which users can create 
and delete accounts, as well as who can modify which accounts.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@167 653ae4dd-d31e-0410-96ef-6bf7bf53c507