Two changes: 1) The default for Posts::get() is now to return all posts, not just the published ones.  If you want your theme to return just the published posts, you'll need to add a status parameter to the get() call.  I suggest that we create some kind of convenience function for this, since I don't like the idea of calling Posts::by_status(Post::STATUS_PUBLISHED) from a theme.  Might just be me.
2) You can now view draft posts that you have written.  This enables certain buttons in the admin for viewing draft posts.  This was enabled by allowing calls to Posts::get() to contain a parameter "where" that can be assigned an associative array of values.  These values are used to construct separate OR'ed where clauses.

For example:

Posts::get(
  'where'=>array(
    array('status'=>Post::STATUS_PUBLISHED),
    array('user_id'=>4)
  )
)

This will return posts that are published OR posts by author_id 4.  Note that the prior syntax of providing any of the where values (status, slug, user_id, or tag) as simple values to the array still works:

// Return published posts only by user #4
Posts::get(
  'status'=>Post::STATUS_PUBLISHED,
  'user_id'=>4
)

Also handy is that you can provide the simple values and the array values so that the simple values carry to each where set:

// Return the post with the slug "foo" that is either
// published or authored by user #4.
Posts::get(
  'slug'=>'foo',
  'where'=>array(
    array('status'=>Post::STATUS_PUBLISHED),
    array('user_id'=>4)
  )
)

And finally, if the 'where' value is a string, it just uses it as the where clause:

Posts::get(
  'slug'=>'foo',  // ignored
  'where'=>"title LIKE '%bar%'"
)

That last one might get extended a bit to use parameterized querying.

I hope that's useful.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@189 653ae4dd-d31e-0410-96ef-6bf7bf53c507