Comments and Plugins
While poking at the comment code to see why Chris couldn't submit new comments, I noticed that the default setting for new comments is "unapproved".  I don't know if this is reasonable or not, but I switched the default to be "approved" for the time being.  And when I did that, I thought, "Hey, wouldn't it be neat if we passed these comments through skippy's plugin code, just to get a rudimentary spam check?"  And so I did.

What we have as a result of banging that code into shape is a single stupid plugin that checks to see if a comment has more than one <a> in it, and if it does, it sets the comment status to Comment::STATUS_SPAM.

Plugins currently don't have a real loading method.  The sample plugin is included explicitly in the main index.php, and it's class is instantiated.  I'm interested in suggestions on how we might store the currently active plugins in the database, and also determine how to list the available plugins from the admin.

From a different standpoint, I would also like to entertain suggestions for a superior comment spam-catching system; Perhaps something karma-related, pluggable, and distributed so that you could use a service like Akismet, but the system would already be so good that you wouldn't bother.

Nonetheless, I'm happy to announce/suggest that the plugins are working, at least to the extent that you can now filter a comment.  (Also, Chris, if you svn up, your site should start accepting comments by default.)

One other change is that I modified the URL class to do two new things.  First, the comment posting URL is now 'slug/"feedback"' instead of just '"feedback"', since we had discussed using a similar idea to reduce the efficacy of stupider spam bots.  We can add something in the future that will allow the "feedback" part to be a hash value seeded by the slug, which could greatly improve this strategy.

Second, the URL class will now default to using incoming settings found in the URL to build new URLs with the ::get() and ::out() methods when those values aren't specifically provided.  This facilitates the above change, and I don't expect it to hurt anything, but report any weirdness.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@190 653ae4dd-d31e-0410-96ef-6bf7bf53c507