Checking some fixes for SQLite compatibilty.
Added a SQLite schema.  This is a bit different from the MySQL schema, 
and we may run into challenges down the road; but it works for now.

Added two new methods to queryrecord:
* exclude_fields
* list_excluded_fields
The first takes a field name (or an array of field names) and adds them 
to the private unsetfields array.  These fields are fields that are 
managed solely by the database, and which Habari should never twiddle.  
The queryrecord::insert method takes the unsetfields and ensures they 
are not passed to DB::insert.  We probably need to add this to 
queryRecord::update, as well.

This all came about because our previous queries were passing the 
auto increment field names to DB::insert, with no value.  MySQL happily 
ignored these, and applied the  column's auto incremenet.  SQLite 
choked on these.  So they had to be excluded.

When making a new object in Habari that relies on a specific database 
table, you'd add the following to your constructor:
   $this->exclude_fields('id');
Assuming id is an auto increment field.

The Post and User objects have been modified to register their excluded 
fields.  Additional fields may be necessary in other objects / classes.

Finally, this commit includes a fix to format.php to for E_ALL 
situations.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@236 653ae4dd-d31e-0410-96ef-6bf7bf53c507