Adding an Error class and applying it to database calls.  Posts now looks for these Error class returns, but other classes should too.
Basically, instead of just returning false when there's a query error, an Error object is returned.  This allows the calling function to process the error and display a result if possible, and also marks the result undeniably as an error rather than just the value "false".

This also allows errors to be logged (a feature that is to be added to the Error class) or sent somewhere or both!

The default error handler is now set to create an instance of Error.  So if there is an unhandled error in the processing of regular code, it is logged.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@250 653ae4dd-d31e-0410-96ef-6bf7bf53c507