I wanted to demonstrate what *exactly* we're planning on doing with community-generated documentation. Compare this to the documentation outline on habariblog.com.
Also, the search system that I started is not complete here.  The idea is that you should be able to load the page directly in a browser (without it being on a web server).  It seems that there are some issues with Ajax calls running locally, so that'll need to be addressed, probably with iframes.  Also, I added a duplicate jquery because I believe that it's important that the help directory be self-contained for this purpose.  (Keep a local copy of just the help.)

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@294 653ae4dd-d31e-0410-96ef-6bf7bf53c507