Prep themes for override class.
The UserThemeHandler class will execute it's Theme instance's methods in place of its own methods in an effort to allow themes to specify their own data gathering routines.  To facilitate this, some of the logic in the constructor for Theme (finding the default theme from the database or using the parameters passed in to create a specific theme) has been moved into Themes::create(), which returns an instance of Theme.  This will allow Themes::create() to return an instance of a Theme descendant for those themes that want to collect more data for display (by implementing CustomTheme::act_display_posts() for example) than what Theme provides.

There is currently not a method to load such a Theme-descendant class, which should be added.

Also included: Moved the logic for selecting a set of posts by date from UserThemeHandler into Posts, so that you can select posts by date via the Posts object API.

All of these changes should be transparent if you're looking only at output, with the exception that day-, month-, and year-based URLs should now return a valid set of posts.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@445 653ae4dd-d31e-0410-96ef-6bf7bf53c507