Updates to the installhandler.php
If a config.php file exists, it is parsed and used to supply the values 
for the database connection portion of the installation screen.
@TODO: must make this SQLite-aware

If a config.php cannot be written, the installer should fail _before_ 
creating any tables.  It's supposed to display the contents of the 
config.php that you need, but it doesn't do that yet.

Removed the option to create a new database from the installer.
Removed extraneous *.tpl files, since we're not using Smarty for the 
installer.
Removed site tagline and about text from installation form.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@457 653ae4dd-d31e-0410-96ef-6bf7bf53c507