Schema changes from issue 211, discussed here: http://groups.google.com/group/habari-dev/browse_thread/thread/24c228d32bd5dc5f
The MySQL and SQLite schemas are both correct for the database types they are associated to.  The SQLite schema is an actual SQLite schema, not a Postgres one.  The Postgres schema is yet to be added as a separate file and tested.

The installer will not currently install on SQLite without some tweaks, which are still being ironed out.

The Post class has been altered slightly to continue to work with additional fields in the posts table that require values that we don't yet provide.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@496 653ae4dd-d31e-0410-96ef-6bf7bf53c507