Installer improvements and beautification.
Added nice HTML for installer, including places for the errors to go.  Will soon be able to verify form details via ajax and expand sections as planned.  Sliced images from graphic design are included with new stylesheet.  The page that displays the config.php when it can't automatically be written has also been updated to display the code output nicely and include a button to return to the main installer page.

Modified InstallHandler to properly install default values to a database when the data supplied in the form doesn't match the data already in the config.  What a pain.  Should see no errors during a MySQL install.

Not tested on SQLite.

Small fix to K2 theme's search page to use MVC $search results instead of requerying.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@497 653ae4dd-d31e-0410-96ef-6bf7bf53c507