More installer fixes.  Of primary interest is restored support for SQLite databases.  I've also shuffled many lines of code from install_db() up to act_begin_install(), since they were not directly responsible for installing the database.  I'm trying to make the code in installhandler easy to follow.
The db_setup.php form has been updated to add a section for the db_file 
for SQLite , but the javascript to show or suppress fields based on the 
DB type selected in the drop-down doesn't seem to be working.  Additionally, 
it'd be nice to be able to automatically select the proper DB type in 
that drop-down based on the value defined in config.php (if any).

If you select to install with SQLite, the installer will check the 
existence of the data file you provide.  If it does not exist, it will 
next check the directory for the file and see if it can be created 
automatically.  If all of that fails, it will return you back to the 
installer.  My tests do not seem to produce any error messages for the 
user, even though the installer is setting error messages with 
$this->theme->assign().

index.php has been updated to check for the existence of the SQLite data 
file, if you're using a SQLite database.

If you want to use a SQLite database, be sure to make the file writable 
to your webserver, or make the directory that will store your datafile 
writable by the webserver.  Then copy config.php.sqlite to config.php, 
and edit it as necessary.



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@502 653ae4dd-d31e-0410-96ef-6bf7bf53c507