Significant updates to the installhandler.
First, I've added a second password field to the install screen, in 
order to make sure folks key in the correct password.

The most significant change in this commit is the inclusion of 
preliminary .htaccess support in the installer.  It's a little rigid 
right now, but mostly functional.

When installing, it first checks whether Habari is being executed on 
Apache.  If it's not, it blithely keeps processing.  We need IIS and 
lighttpd rewrite support in here.

Next it checks for the presence of a .htaccess file.  If it does not 
exist, it attempts to create one.  If that fails, installation is 
prevented until you manually construct the .htaccess file (using the 
template that the installer displays).

If .htaccess exists but does not contain the string "HABARI" (a lame 
check, but sufficient for now), the installer attempts to add the 
necessary bits.  Again, if it can't write it will show you what you need 
to manually insert.

The unfinished functionality is updating an .htaccess file that contains 
some but not all of the Habari rules without trouncing any existing 
rules.



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@536 653ae4dd-d31e-0410-96ef-6bf7bf53c507