Introduces autop functionality to feed output to supercede changes in r534.
Added a constructor to AtomHandler that applies format filters to the Atom output only when it is requested when not using a login.  If you are logged in and request the Atom entry collection, you will not received autop-filtered entries.  This is because APP uses the entry collection for raw data when editing.  If you're editing a post, you should edit the raw data, not filtered data.

For users who are not logged in, the filters will be applied.  

This code should probably be applied from a filter that can be removed easily, such as in a default plugin.  It should possibly also have separate URLs for feed readers versus APP clients.

Additional notes in the source.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@537 653ae4dd-d31e-0410-96ef-6bf7bf53c507