Minor update to htdocs/system/admin/post.php to programmatically present the list of available post statuses.
We need UI to handle the display of private posts.  Perhaps a selection 
list on content.php to filter which kind of content to display?

Also included in this commit is a minor re-ordering of the 
list_post_types() and list_post_statuses() logic, so that the 'any' 
name/value pair is the first in the list, and values returned from the 
database are presented in ascending order by ID.



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@558 653ae4dd-d31e-0410-96ef-6bf7bf53c507