Radio buttons should now work like radio buttons on comment moderation pages with this commit.
Modified the Comment class to correctly reference tables and delete the comment to which the instance points, rather than any specific id (which is better done through Comments::delete_these()).

Modified the Comments class to indicate static versus dynamic functions.  Note that the previously named Comments::delete() is now Comments::delete_these(), a static function.  This is because an instance of the Comments class represents a set of comments.  If you call $comments->delete(), you should expect it to do the same thing as $comment->delete(), which is to delete the comments held by the instance.  In other words, while the static function Comments::delete_these() accepts a parameter of which comments to delete, Comments::delete() should not accept a parameter and should delete the comments already held in the instance.  Comments::delete() does not exist in this commit.

Chris' system/admin/comments.php is included in this commit with changes to the radio button style.  This page is not currently referenced in the menu, but is accessible via domain.com/admin/comments .  We should discuss whittling down the number of comment moderation pages so that it's perhaps less confusing which page you need to look at to get the information you want.  Maybe an Ajax-powered switch at the top of the page?

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@561 653ae4dd-d31e-0410-96ef-6bf7bf53c507