This fixes a couple of small issues.
First, the login form is prefixes with "habari_", and userhandler.php is 
updated to deal with it accordingly.  This address #46.  More review 
will be necessary to determine if additional form fields require this.

This also fixes pagination of search results, as described in issue 
#264.  Thanks vkaryl for pointing it out, and freakerz for the bulk of 
the fix.  The page indicator is still not updating correctly when paging 
through search results.  See comment #11 on issue #264.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@569 653ae4dd-d31e-0410-96ef-6bf7bf53c507