The update beacon, first pass. A call to Update::check() returns an array of data for components that are known to need an update.  This is done on the dashboard and output is produced.
Three types of update are imagined:  bugfix, feature, critical.  The server will only report on updates that have taken place since the last critical update, including that critical update.  The Habari installation figures out (in the Update class) which updates apply to the current install.  This will save some computational wear on the update server.

Note that the beacon at http://beacon.habariproject.org is currently a static file that returns information that makes Habari think that it needs a critical update.

Also note that the beacon is requested on every view of the dashboard.  This should be run only periodically.  The beacon data will need to be cached and compared to the requests for updates.  If the installed components change, then the request to the beacon should fire immediately instead of using cached data.  This paragraph represents planned but currently unimplemented functionality.

Plugins or themes that wish to check the central beacon for updates should register the update_beacon action, and call Update::add() to insert their beacon into the update check array.

A "beacon id" is a UUID that represents the component that requires updating.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@583 653ae4dd-d31e-0410-96ef-6bf7bf53c507