Updates to post deletion logic.  Moved the initial status change from adminhandler to post::delete().  In this way, one need only ever call $post->delete() to have that post's status set to "deleted".  Subsequent calls to $post->delete() will purge the post from the database.
Post::delete() also accepts an optional boolean argument.  This 
argument, if true, will immediately delete the post.  The defautl value 
for this option is false.

Implemented plugin hooks for post deletion, as well as changes to a 
post objections member properties.  post_delete actions will be 
triggered when deleting a post; and post_update_* will be triggered when 
making changes to a post object in Post::update().



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@589 653ae4dd-d31e-0410-96ef-6bf7bf53c507