Modified InfoRecords to cache info metadata until calling $info->commit();
This allows Post and Comment objects to have their info set before they are saved.  When they are saved, they get an ID, and the ID is passed to the commit() function in the InfoRecord.  This is useful, for example, for setting info metadata while checking a comment for spam from a plugin.

Only changed values in info are updated on commit().

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@600 653ae4dd-d31e-0410-96ef-6bf7bf53c507