Added hook to Post to allow post data to be passed in whole to plugin sink.
Added Format::apply_with_hook_params() to Format class for applying formatters that also accept all plugin hook data as parameters, not just the first parameter.

Added Format::more(), which when used with apply_with_hook_params() allows a theme to display the first part of a post (prior to <!--more--> in the post content) on pages displaying more than one post, and the second part of a post only on pages that display that post alone.


git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@655 653ae4dd-d31e-0410-96ef-6bf7bf53c507