ACL. Added a fixed-up derivative of skippy's ACL class.  Doesn't currently have methods to grant or deny permissions, but will successfully cache any existing permissions from the database and return them if you call $user->can('action'). Rebuilt __autoload() to cache files instead of file_exists()'ing them each time.  May be expanded in the future to allow subdirectories inside of /classes for the sake of organization.  This also allows the new __static() method to exist within a class.  __autoload() will call this static method on a class if it exists as soon as it is loaded, avoiding filling dozens of static class methods with if(!isset(self::$foo)) {self::setfoo();} Added required 'admin' permission to the entire adminhandler class for testing purposes. Added four new permissions/groups tables to DatabaseConnection. Removed the karma stuff from the user class and replaced it with the ->can() method. Bumped the DB and API version to trigger the database upgrade. Updated both the MySQL and SQLite schemas to include the four new tables.
No permissions are added to the database by default.  The ACL class contains a constant that defines what to do if a permission is not defined.  Currently, it grants permission on any request for a permission that does not exist.  The end result is that if you upgrade, you shouldn't really notice anything.  But if you add the permission 'admin' to the permissions table, then you will be locked out of your admin until you put data into the other three tables that actually grants you that permission.  The side effect is that you can disable any permission in the system by removing it from the permissions table, which I think is kind of cool, but should be discussed.

This commit might break multisite.  I was having trouble with multisite after the change to __autoload() and although I think the issue is in the Site class, it's entirely possible (probable) that it is related to the __autoload() changes.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@743 653ae4dd-d31e-0410-96ef-6bf7bf53c507