Added 'template_fallback' filter, for adding additional fallback templates to the list from a plugin. Added 'template_user_filters' filter, for adding additional parameters to the Posts::get() request done in Theme::act_display() from a plugin.  These two filters are courtesy of freakerz, on issue #368, although I appended "template_" to the filter names for improved readability.
Also, the UserThemeHandler no longer prevents __call() overrides on Theme, so it's now possible to include for execution a Theme::__call() method, which is now present.  Currently, this method will execute the 'theme_action' plugin action for any theme method that is not defined by the current theme.  For example, if you add a new rewrite rule that dispatches to UserThemeHandler::kumquat(), the kumquat() method will likely not exist in the Theme class.  Habari will then call Plugins::act('kumquat', $theme, array()) to handle that request.  In essence, this allows plugins to define functions that implement new theme methods.

Let's see if this works.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@751 653ae4dd-d31e-0410-96ef-6bf7bf53c507