This commit removes undelete functionality from the core of Habari. After updating to this revision, deleting a post will immediately remove that post from the database.
The Post class has been updated to provide the following methods:
* add_new_type -- register a new Post type to the posttype table
* add_new_status -- register a new Post status to the poststatus table

The Post class has also been modified to consistently invoke Plugins:
* allow_post_* -- a filter used to determine if a specific action
(insert, update, delete) should be permitted to execute
* before_post_* -- an action that is called prior to something happening
(insert, update, delete)
* after_post_* -- an action that is called after something happens
(insert, update, delete)

Included is the first stab at a new Undelete plugin.  It provides a
filter_allow_post_delete method which interrupts the post deletion
process.  It captures the post's current status in a postinfo record,
then sets that post's status to "deleted".  When the plugin is
activated, it will register the "deleted" status in the poststatus
table, if it doesn't already exist.
This plugin still needs some work:
* admin interface to allow restoring deleted posts to their prior status
* add CSS elements to the current theme, to re-address issue #354.

This commit adds the $post->statusname property, per issue #372.  
Modified /user/themes/k2 template files to make use of this.

This commit adds a few index.html files, to prevent directory listings.
It's not real security, but it's a fair step to make it harder to
determine what plugins are installed on the site.

This commit modifies the SQL schema files to remove some INSERT 
statements.  Core Habari API calls are used instead to register EventLog 
types, as well as default post types and statuses.  Closes issue #374.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@761 653ae4dd-d31e-0410-96ef-6bf7bf53c507