First stab at standardizing pluign hook names, in the form         class_method_extra
For example, the hooks that execute before and after a post is deleted are:
        post_delete_before
        post_delete_after

I've only touched the Post and Comment classes for now.  Also included is a filt
er:
        <class>_<method>_allow
This will allow plugins to determine whether a post/comment may be inserted/upda
ted/deleted.

I've modified the core plugins to work with the new hook names.  I've also modif
ied the feedbackhandler class to not invoke add_comment as a hook.  It still inv
okes spam_filter.

Finally, I've upated the author value on the core plugins to "Habari Community", per ticket 379.



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@771 653ae4dd-d31e-0410-96ef-6bf7bf53c507