Updated the APP namespace.
Thanks to ringmaster for pointing out the update of APE (Atom Publishing Exerciser). It now uses the latest draft-17 which according to Tim Bray, maker of APE and member of Atom Workgroup, is the final draft to be used as the newest RFC by IETF.

Therefore, I'll be progressively updating the AtomHandler to try and support as many features as possible. Hopefully this means desktop clients will follow suit and work with Habari.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@803 653ae4dd-d31e-0410-96ef-6bf7bf53c507