Fixed the pagination problem for good, hopefully.
If you omit to supply the proper settings, the function will now get them from the handler_vars which contains _GET, _POST and the URL matches.

The RewriteRule::match() does not fill ->named_args_values with the _GET and _POST matchings, so it caused the search criteria to not be passed along the pagination.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@859 653ae4dd-d31e-0410-96ef-6bf7bf53c507