Fixing an issue with the installer, if config.php was not writable, the write_config_file() would use $db_host and not $db_file, therefore the file comparison would fail and you could not install Habari.
Now it has a switch(), because if we add support for other database engines, we'll need to add more cases for each of them since they will probably use different DSN strings.

I also removed a check on $_POST['db_user'], that is not used by SQLite, we have several other checks prior and after for these kind of errors.

Briefly, this means we still have MySQL-specific code lurking!

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@946 653ae4dd-d31e-0410-96ef-6bf7bf53c507