Added Post Type De/Activate
Fixes #434

* Updates the schema for both MySQL and SQLite to add an active column to posttype table (tinyint(1))
* Changes the existing list_post_types function to list_active_post_types and returns only posts with active == 1
* Adds a function list_all_post_types, which returns each post type with its id and active value as an array
* Updates all files referencing Post::list_post_types to Post::list_active_post_types as thats what they are expecting
* activate_post_type()
* deactivate_post_type()

I also added an argument to add_new_type, $active.  This is the value that goes into the DB, so we can add post types but mark them inactive at the time of creation.

This means we have the ability to add post_types (marking them in/active on creation), we can then de/activate them easily.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@960 653ae4dd-d31e-0410-96ef-6bf7bf53c507