Modified QueryRecord::__get() to work like Controller::get_var().
If none is set, return NULL. This way you don't get a unexpected exception when asking for $comment->gravatar, for example, since we know there's a plugin to handle that call.

This works for $post too, $post->gravatar would return NULL.

I'ld like to use an other way, but for now that's all I can come up with, and since Controller::get_var() uses that, I thought of just copying that behavior.. it's somewhat logic anyways.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@980 653ae4dd-d31e-0410-96ef-6bf7bf53c507