Added admin message handling to the Session class.
To create a message for display use either Session::notice() or Session::error().  Use Session::messages_out() in the admin to output the messages in a nice ul.  The messages are stored in the user's session, and so persist between page loads.  End result:  You can specify much more information than you could by passing statuses around in the URL.

Look at the admin publish.php and AdminHandler for an example of this in use.  The status AND name of the post appear in the update box when you save a post.

We need to consider using theme functions with passable parameters to output stuf like this, so that, for example, we could pass the message arrays to a template to have them formatted by the theme rather than formatting them in code.  More thought and code is required here.

Also tried to fix a glitch in dashboard.php where the message property of the exception was incorrectly fetched.

Happy Thanksgiving.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@1120 653ae4dd-d31e-0410-96ef-6bf7bf53c507