Checking in work-in-progress on usergroup stuff.
Although no menu link is yet provided, one may manually navigate to 
/admin/groups to see the beginnings of UserGroup management.  One may 
create and delete groups, as well as add and remove members from those 
groups.

The login in adminhandler::groups() is inelegant, and could use some 
attention.  Likewise, the groups.php admin file is less than ideal.  But 
this is a start.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@1286 653ae4dd-d31e-0410-96ef-6bf7bf53c507