*MUST READ, THIS WILL BREAK YOUR PAGINATION*
Moved Utils::page_selector() to Theme::, meaning you can now call it via:
$theme->page_selector()

Also, moved the auto-detection of RewriteRule arguments in URL::get(), if you supply NULL as the rule name, it will use the current matched rule and get the arguments automagically.

In the end, look at the change log, it's quite easy to make out the changes!

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@1574 653ae4dd-d31e-0410-96ef-6bf7bf53c507