M    post.php
Major fixup for save_tags() which actually loops through the posted and
existing tag slugs and text values and determines what should be inserted,
what should be deleted, and what should be mapped or unmapped in the tag2post
table.

M    adminhandler.php

A very bad usability bug fixed.  It seems that if the __set() magic function is
called with an empty value, it is ignored.  Therefore, when Post::get_by_slug() is
called, it populates $post->tags with the original set of tags.  However, when deleting
all tags from a post, an empty array is passed to $post->tags= $tags; which is converted
to Post::__set('tag', array()) and subsequently ignored.  This means the original tags
array is passed to save_tags() and the mapping records were not being deleted correctly.


git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@1781 653ae4dd-d31e-0410-96ef-6bf7bf53c507