Added some of Matt Read's changes to the Modules class (basically everything that was compatible with the goal of identifying modules by a unique ID rather than by name).
Implemented AJAX adding and removal of dashboard modules. We have not settled
on a UI for removal of modules. I have included one example of how this could
be done with the latest log module. Another option would be to have a close
button on the front of the modules.

I will be adding the ability to filter the logs module display with config
options, so please don't remove that module's backpanel.

Renamed Latest Posts to Latest Entries, because it really only shows posts
of the 'entry' type.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@1992 653ae4dd-d31e-0410-96ef-6bf7bf53c507