Beginning work on the options page. The page is now completely generated by formUI. All option 'items' are listed in the dropdown box at the top-left of the page, and selecting an option here filters the available options. (Implementation of the search box coming soon.)
Some notes about these changes: I have added controls for all the options
which were listed on the existing options page. Some of these controls might
not actually do anything, but formUI will save even the empty options to the
habari__options table.

There has been discussion on IRC about having all plugin options be on the
options page. With the new structure of the page, it will be much easier to
accomplish this, but it is not yet implemented.

The previous version of the page had 'help text' for some of the options. This
text is available in the $option_items array but I am not currently doing
anything with it.


git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@2065 653ae4dd-d31e-0410-96ef-6bf7bf53c507