Adding completed translations from https://translations.launchpad.net/habari/trunk/+pots/habari/.
As you can see the "POT" (template) file is out of date, we should make 
sure we update this before 0.5 is released to give the translation teams
time to update their translations.

We should also have in place a mechanism to automatically generate the 
pot file and have it kept up-to-date with trunk.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@2156 653ae4dd-d31e-0410-96ef-6bf7bf53c507