Adding a generic System Status dashboard module.
If a plugin hooks filter_dashboard_status($statuses) and adds a keyed value to the $statuses array, it will cause the System Status module to appear with that value. 

This is useful for modules that just want to display a simple value or alert in the dashboard but don't want to add a whole module.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@2630 653ae4dd-d31e-0410-96ef-6bf7bf53c507