Adds a not:id criteria to filter out specific post IDs that aren't wanted.
Also, in testing we noticed that there can be ambiguous field names in the query, so this attempts to quickly prefix all field names with their {table}s. I haven't tested this thoroughly and chances are good that I've missed something or typo'd something, so be on the lookout. Also, this only covers the first criteria portion of the method - the last half needs to have table prefixes added too.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@3343 653ae4dd-d31e-0410-96ef-6bf7bf53c507