Create an upgrade path for SQLite from 0.5.2 to 0.6. Simplify the SQLite upgrade script. Remove upgrade scripts that allowed interim updates for individuals uses the svn version of habari, but aren't needed when going from 0.5.2 to 0.6.
The use of PRAGMA syncronous = OFF; in SQLite's connection function results in major performance increases at the expense of a possible corrupt db if the server OS crashes or loses power. The default for this setting is 'FULL', which is very safe, but very slow. There is also a middle setting, 'NORMAL', which we may want to consider. It is between the other two settings in both safety and performance.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@3423 653ae4dd-d31e-0410-96ef-6bf7bf53c507