Modifying the operation (but hopefully not the existing behavior) of Posts::get() for info fields.
This change allows you to use the originally present all:info, any:info, and has:info parameters to Posts::get(), but now uses a LEFT JOIN for each info name instead of a subquery.  The result is that the info record values are now included in the query results and can be used in the orderby parameter.

For example, to fetch posts that have a postinfo value named "beta" and sort on it:
$posts = Posts::get(array('has:info'=>'beta', 'orderby' => 'info_beta_value ASC'));

The "not:" operands of the info parameters are unaltered, but should probably also use this method.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@3855 653ae4dd-d31e-0410-96ef-6bf7bf53c507