Check permissions before going to the create publish page, before actually creating a new post, and if the user tries to change a post's content type. Props to Luke @Squareweave for the original patch. Modified to add checks in the main menu, and explicit checks for denied access to the specified content type. Fixes ticket #1075.
There is a change of behavior here in that the user has to have either permission to create the specified type, or create permission on posts_any, rather than create permissions on own_posts. I'm leaning toward the idea that create permissions on specified content types should be the only applicable ones when it comes to giving permission to create a post, but there it is.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@3904 653ae4dd-d31e-0410-96ef-6bf7bf53c507