Make Posts::get() capable of dealing with multiple vocabularies by extending the colon syntax we currently use for tags. This may not be the best way, but it involves relatively few changes and keeps the parameter syntax for Posts::get() reasonable. AtomHandler, Theme, and the search box on the manage posts page in the admin have been updated to deal with the change. Normal users shouldn't see any difference in behavior. Tags still live at www.site.com/tag/tagname, the atom feed still lives at www.site.com/tag/tagname/atom, and the search box on the manage posts page still uses tag:
Any plugins or themes using custom calls to Posts::get() using the variations on the tags parameter will have to be updated. A new parameter, 'vocabulary', has been added. The syntax is 

Posts::get( 'vocabulary' => array( 'tags:term' => 'habari', 'tags:not:term' => 'asides', 'tags:term_display' => 'Community', 'tags:not:term_display' => 'House' ) );

Note that the vocabulary parameter is an array of parameters. Here, 'tags' is where the vocabulary name is prepended. 'term' is the equivalent of 'tag_slug', 'term_display' is the equivalent of 'tag'.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4380 653ae4dd-d31e-0410-96ef-6bf7bf53c507