Refactor Options::get_group() so that it just wraps around Options::get(). If we access the array of values directly we don't give Plugins the chance to filter them.
Add Options::set_group() that is the reverse of get_group(). Specify a prefix and an associative array of names => values and we'll loop through setting each one.

Add Options::delete_group() that behaves exactly as you would expect.

Add the ability to pass an array of name => value pairs to Options::set() as the first parameter and set all of them.

Add the ability to pass an array of names to Options::delete() and delete all of them.

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4397 653ae4dd-d31e-0410-96ef-6bf7bf53c507