Add block renaming and more flexible config.  Blocks should not output their own titles from now on, but allow the theme to provide a wrapper for the titles instead.
Update FormUI to allow you to specify an array as a storage method.  The first element of the array is an action hook name.  Values are passed to that hook to affect storage of that item.

All blocks now display a title textbox and a "show title" checkbox in the config modal dialog.

Updating the block name in the config updates the block name on the parent form via ajax.



git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@4441 653ae4dd-d31e-0410-96ef-6bf7bf53c507