Reverting the changes introduced in r3605 and refactored in r3621 which introduced a second session notice on publishing of new content indicating the URL slug of the post.
This information is available from both the 'View Post' link as well as the 'Content Address' field under the Settings tab. There are numerous reasons the content URL would not match the exact title of the post you created (duplicate titles, non-URL characters, encoding differences, plugin modifications, custom rewrite rules, etc.). You should look at the content address field or copy-and-paste the View Post link instead of manually anticipating what the slug will be. Displaying a session notice is redundant and intrusive and only displayed the post slug, not the entire URL that would be needed to reach the post, making it further error prone.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@4454 653ae4dd-d31e-0410-96ef-6bf7bf53c507