Rewrite the 'vocabulary' section of Posts::get(). The syntax from a few weeks ago should still work. In addition, 'id' is now a valid parameter, so you can do something like
'vocabulary' => array( 'term:id' => $tag->id)

Additionally, there is now a simplified syntax allowing you to pass in tag and term objects. The syntax is

'vocabulary' => array( 'any' => array($tag1, $tag2), 'all' => array( $tag3, $tag4), 'not' => $tag5)

not: this term is not allowed
any: these terms get ORd together. any of them can be on the posts
all: these terms get ANDed. all of them must be on the posts

Note that this means you don't have to pass in the vocabulary if term objects are passed in. The code takes care of it for you.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@4512 653ae4dd-d31e-0410-96ef-6bf7bf53c507