Add a $force option to plugin deactivation, which forces a plugin to be deactivated, regardless of what other plugins say.
When loading active plugins, if the file does not exist, forcibly deactivate the plugin. If you don't, a broken plugin could randomly pop back up as active if you move it back to its original location. I'm wiling to bet we all have old plugins in the list that we wouldn't want to magically reappear as soon as their files show back up.

git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@4611 653ae4dd-d31e-0410-96ef-6bf7bf53c507