Rework the interface for adding blocks to areas, trying to reuse interface elements from the rest of the admin.
Click the cross to remove a block from an area, the same as you would to remove a module from the dashboard. Add blocks to areas via the dropbutton. Future work includes adding a handle to the draggable blocks (as per the modules on the dashboard) and putting 'No blocks' in areas that have no blocks added yet. Playing with the labels, such as adding "Available blocks" and "Areas in scope" might make things clearer. The javascript and css also needs a clean up.

Let me know if I've inadvertently removed functionality. This is committed in the spirit of commit then review.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@5000 653ae4dd-d31e-0410-96ef-6bf7bf53c507