habari.media.resize_media_row now resizes the media_row of the visible silo tab using information from the visible silo and not the first silo in the list. This fixed #1403.
The problem here was the resize_media_row was using the $('.mediaphotos').outerWidth() of the first "mediaphotos" it found and not the one of the active silo tab.  Accordingly, the result of this little calculation was always 0 when NOT viewing the first tab and hence the row was never resized and we didn't see the silo contents on Chrome.  

I'm not exactly sure why this was only a problem on Chrome, but I suspect it may have something to do with the fact Chrome seems to enforce the "display:block" for divs in the default user agent stylesheet differently from Firefox.

This fix may be a bit crude, but it does the job, and consistently across Chrome, Safari and Firefox on OS X.  Being more silo specific should probably be considered elsewhere in media.js to reduce the chances of a similar issue occurring in the future, but I'll leave that for the moment.


git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@5113 653ae4dd-d31e-0410-96ef-6bf7bf53c507