Allow Charcoal to correctly use multiple plural forms when the language requires them.
While you won't be able to tell because the strings haven't been translated yet, _n() needs to have the original text to pick the proper plural form, not the translated text. Also, as one of the comments on the ticket noted, put the number format string in the translation file so the translator can pick the proper order of it and the plural. Fixes habari/habari#41