Replace svn checks with similar git functions. Git does not do string replacements, so the check for whether Habari is a development version is considerably less sophisticated.
Which is to say, it checks the string for a hyphen as would be seen in "0.9-alpha" or "0.9-rc1", etc.

This is not necessarily a good way to do this, but since the is_devel() check now only happens on the sysinfo page, the impact should be considerably lessened.

Bump API_VERSION.

Closes habari/habari#291.