Add a 'class' parameter for menu items.
Add CSS for menu splitters, and update the main menu to use them in place of the existing hardcoded ones.

It is possible now to create a splitter by adding class "over-splitter" to the item above it, and "under-splitter" to the item below.

A single CSS class could probably replace the two of them.

Closes habari/habari#314 and habari/habari#315