Fix HTTPS detection, again.
I knew there was a reason habari/habari#288 would break things, I just couldn't remember it.

OpenStack sets HTTPS to 1, which is not equal to 'on'. So let's flip the logic back since we have a fairly inclusive list of things which indicate HTTPS is not being used and allow for just about anything else to indicate it is.

Our current list of possible conditions: off, on, empty string, false, null, and 0 / 1. There really should be a unit test for this...