Fix handling of manually-specified published dates (or lack thereof) when a post is published.
This may fix #395.

Essentially, the prior change in 3fc4e64a seems to have been a bit of a typo. We don't want to compare `pubdate` and `updated`, we want to compare `pubdate` with `pubdate`. The key change in that commit was merely the ==, rather than >= -- we don't care if the date they specify is in the past or future, simply that it's different.

So to clear everything up: When you publish a post (when the state goes from something that is not published to published, whether for the first or hundredth time) if you have not manually specified a publish date on the form, the publish date will be updated to the current time.