Publish scheduled posts if their dates are changed into the past.
This is the opposite of the logic which changes a published post to scheduled if the date is in the future.

This could happen if you publish a post with a date in the future (so it becomes scheduled), then edit again and move the date into the past and save -- it would remain 'scheduled', when it should be made 'published'.