When reading sessions, don't delete anything.
Instead, we'll just not let you use the session if it's not valid and expire your session cookie. If you then write to the session, a new one will get generated. Otherwise, you're back out in the cold, as if you hadn't pretended to have one in the first place.