Prevent backslash from tripping logging with PostgreSQL
The "data" field is of type BYTEA (BLOB in other databases). It is observed
that PHP is flawed with interpreting PARAM_LOB, it accepts/returns a file stream
for one database connector (pgsql) and a string for others (sqlite, mysql?).
Since modifying the whole API just to satisfy a single data type addition is
a bit overkill, drop that idea and patch the string for postgresql such that it
is not possible to die on data like `Habari\Method`. (escape sequences for
PostgreSQL are documented for bytea[1]).

 [1]: http://www.postgresql.org/docs/devel/static/datatype-binary.html