Revert "SECURITY FIX. Related to the one before. Stop users from deleting posts when they don't actually have delete access to that post."
The original commit was a bit overzealous.  There are many places in
system that cease to function properly when access control is checked on
every commit, and in particular, this commit causes all extant posts tests
to fail.  Instead, access control should be checked by the thing that's
calling the delete, not the delete function itself.

This reverts commit 9c57ef3ac9a203ddb8cb03cb6f85376739c292a9.