Display the site title on the login page.
Having a generic login form means that for most themes it's not necessary to
include a login template, instead falling back to the generic template.
Displaying the Habari logo so prominently means that the login template is no
longer generic, making it less attractive as a fallback. Additionally, while
the logo looks good here, the default login page no longer identified the site
you're logging into, which is a backwards step from a usability point of view.

Don't hate me, @chrisjdavis, let's talk about it if you feel strongly.